Use SPYDER_DEV env var to *not* rename dot config files when using bootstrap.py
- This change is meant *only* for developers. It'll force them to maintain
separate settings for 2.2 and 2.3dev.
- The idea is to make easier to switch between these branches without losing or
overwriting settings.
- The intended audience is those of us who develop Spyder with Spyder.

+++ spyderlib/userconfig.py
@@ -207,6 +207,7 @@ class UserConfig(cp.ConfigParser):
         """
         folder = get_home_dir()
         w_dot = osp.join(folder, '.%s.ini' % self.name)
+        DEV = os.environ.get('SPYDER_DEV')
         if self.subfolder is None:
             return w_dot
         else:
@@ -217,7 +218,7 @@ class UserConfig(cp.ConfigParser):
                 # Folder (or one of its parents) already exists
                 pass
             old, new = w_dot, osp.join(folder, '%s.ini' % self.name)
-            if osp.isfile(old):
+            if osp.isfile(old) and not DEV:
                 try:
                     if osp.isfile(new):
                         os.remove(old)
